Make safe requests to scheme://region/bucket.
Making requests to scheme://bucket.region only works for a subset of S3 buckets. It assumes the S3 bucket name is a valid hostname and someone may have an underscore in the name. It also causes requests to S3 buckets that have full stops in the name since the SSL certificate will not be valid for the subdomain of the bucket subdomain.